+++ tiledb/sm/filesystem/posix.cc
@@ -1,5 +1,5 @@
 /**
- * @file   posix_filesystem.cc
+ * @file   posix.cc
  *
  * @section LICENSE
  *
@@ -32,9 +32,10 @@
 
 #ifndef _WIN32
 
-#include "tiledb/sm/filesystem/posix_filesystem.h"
+#include "tiledb/sm/filesystem/posix.h"
 #include "tiledb/sm/misc/constants.h"
 #include "tiledb/sm/misc/logger.h"
+#include "tiledb/sm/misc/stats.h"
 #include "tiledb/sm/misc/utils.h"
 
 #include <dirent.h>
@@ -49,22 +50,12 @@
 namespace tiledb {
 namespace sm {
 
-namespace posix {
-
-bool both_slashes(char a, char b) {
+bool Posix::both_slashes(char a, char b) {
   return a == '/' && b == '/';
 }
 
-/**
- * Reads all nbytes from the given file descriptor, retrying as necessary.
- *
- * @param fd Open file descriptor to read from
- * @param buffer Buffer to hold read data
- * @param nbytes Number of bytes to read
- * @param offset Offset in file to start reading from.
- * @return Number of bytes actually read (< nbytes on error).
- */
-uint64_t read_all(int fd, void* buffer, uint64_t nbytes, uint64_t offset) {
+uint64_t Posix::read_all(
+    int fd, void* buffer, uint64_t nbytes, uint64_t offset) {
   auto bytes = reinterpret_cast<char*>(buffer);
   uint64_t nread = 0;
   do {
@@ -82,19 +73,13 @@ uint64_t read_all(int fd, void* buffer, uint64_t nbytes, uint64_t offset) {
   return nread;
 }
 
-/**
- * Writes all nbytes to the given file descriptor, retrying as necessary.
- *
- * @param fd Open file descriptor to write to
- * @param buffer Buffer with data to write
- * @param nbytes Number of bytes to write
- * @return Number of bytes actually written (< nbytes on error).
- */
-uint64_t write_all(int fd, const void* buffer, uint64_t nbytes) {
+uint64_t Posix::pwrite_all(
+    int fd, uint64_t file_offset, const void* buffer, uint64_t nbytes) {
   auto bytes = reinterpret_cast<const char*>(buffer);
   uint64_t written = 0;
   do {
-    ssize_t actual_written = ::write(fd, bytes + written, nbytes - written);
+    ssize_t actual_written =
+        ::pwrite(fd, bytes + written, nbytes - written, file_offset + written);
     if (actual_written == -1) {
       LOG_STATUS(
           Status::Error(std::string("POSIX write error: ") + strerror(errno)));
@@ -107,7 +92,7 @@ uint64_t write_all(int fd, const void* buffer, uint64_t nbytes) {
   return written;
 }
 
-void adjacent_slashes_dedup(std::string* path) {
+void Posix::adjacent_slashes_dedup(std::string* path) {
   assert(utils::starts_with(*path, "file://"));
   path->erase(
       std::unique(
@@ -117,7 +102,7 @@ void adjacent_slashes_dedup(std::string* path) {
       path->end());
 }
 
-std::string abs_path(const std::string& path) {
+std::string Posix::abs_path(const std::string& path) {
   // Initialize current, home and root
   std::string current = current_dir();
   auto env_home_ptr = getenv("HOME");
@@ -152,9 +137,9 @@ std::string abs_path(const std::string& path) {
   return ret_dir;
 }
 
-Status create_dir(const std::string& path) {
+Status Posix::create_dir(const std::string& path) const {
   // If the directory does not exist, create it
-  if (posix::is_dir(path)) {
+  if (is_dir(path)) {
     return LOG_STATUS(Status::IOError(
         std::string("Cannot create directory '") + path +
         "'; Directory already exists"));
@@ -167,7 +152,7 @@ Status create_dir(const std::string& path) {
   return Status::Ok();
 }
 
-Status touch(const std::string& filename) {
+Status Posix::touch(const std::string& filename) const {
   int fd = ::open(filename.c_str(), O_WRONLY | O_CREAT | O_SYNC, S_IRWXU);
   if (fd == -1 || ::close(fd) != 0) {
     return LOG_STATUS(Status::IOError(
@@ -178,7 +163,7 @@ Status touch(const std::string& filename) {
   return Status::Ok();
 }
 
-std::string current_dir() {
+std::string Posix::current_dir() {
   std::string dir;
   char* path = getcwd(nullptr, 0);
   if (path != nullptr) {
@@ -190,7 +175,7 @@ std::string current_dir() {
 
 // TODO: it maybe better to use unlinkat for deeply nested recursive directories
 // but the path name length limit in TileDB may make this unnecessary
-int unlink_cb(
+int Posix::unlink_cb(
     const char* fpath,
     const struct stat* sb,
     int typeflag,
@@ -204,7 +189,7 @@ int unlink_cb(
   return rc;
 }
 
-Status remove_dir(const std::string& path) {
+Status Posix::remove_dir(const std::string& path) const {
   int rc = nftw(path.c_str(), unlink_cb, 64, FTW_DEPTH | FTW_PHYS);
   if (rc)
     return LOG_STATUS(Status::IOError(
@@ -213,7 +198,7 @@ Status remove_dir(const std::string& path) {
   return Status::Ok();
 }
 
-Status remove_file(const std::string& path) {
+Status Posix::remove_file(const std::string& path) const {
   if (remove(path.c_str()) != 0) {
     return LOG_STATUS(Status::IOError(
         std::string("Cannot delete file '") + path + "'; " + strerror(errno)));
@@ -221,7 +206,7 @@ Status remove_file(const std::string& path) {
   return Status::Ok();
 }
 
-Status file_size(const std::string& path, uint64_t* size) {
+Status Posix::file_size(const std::string& path, uint64_t* size) const {
   int fd = open(path.c_str(), O_RDONLY);
   if (fd == -1) {
     return LOG_STATUS(Status::IOError(
@@ -236,7 +221,8 @@ Status file_size(const std::string& path, uint64_t* size) {
   return Status::Ok();
 }
 
-Status filelock_lock(const std::string& filename, filelock_t* fd, bool shared) {
+Status Posix::filelock_lock(
+    const std::string& filename, filelock_t* fd, bool shared) const {
   // Prepare the flock struct
   struct flock fl;
   memset(&fl, 0, sizeof(struct flock));
@@ -263,26 +249,40 @@ Status filelock_lock(const std::string& filename, filelock_t* fd, bool shared) {
   return Status::Ok();
 }
 
-Status filelock_unlock(filelock_t fd) {
+Status Posix::filelock_unlock(filelock_t fd) const {
   if (::close(fd) == -1)
     return LOG_STATUS(Status::IOError(
         std::string("Cannot unlock filelock: ") + strerror(errno)));
   return Status::Ok();
 }
 
-bool is_dir(const std::string& path) {
+Status Posix::init(
+    const Config::VFSParams& vfs_params, ThreadPool* vfs_thread_pool) {
+  if (vfs_thread_pool == nullptr) {
+    return LOG_STATUS(
+        Status::VFSError("Cannot initialize with null thread pool"));
+  }
+
+  vfs_params_ = vfs_params;
+  vfs_thread_pool_ = vfs_thread_pool;
+
+  return Status::Ok();
+}
+
+bool Posix::is_dir(const std::string& path) const {
   struct stat st;
   memset(&st, 0, sizeof(struct stat));
   return stat(path.c_str(), &st) == 0 && S_ISDIR(st.st_mode);
 }
 
-bool is_file(const std::string& path) {
+bool Posix::is_file(const std::string& path) const {
   struct stat st;
   memset(&st, 0, sizeof(struct stat));
   return (stat(path.c_str(), &st) == 0) && !S_ISDIR(st.st_mode);
 }
 
-Status ls(const std::string& path, std::vector<std::string>* paths) {
+Status Posix::ls(
+    const std::string& path, std::vector<std::string>* paths) const {
   struct dirent* next_path = nullptr;
   DIR* dir = opendir(path.c_str());
   if (dir == nullptr) {
@@ -302,7 +302,8 @@ Status ls(const std::string& path, std::vector<std::string>* paths) {
   return Status::Ok();
 }
 
-Status move_path(const std::string& old_path, const std::string& new_path) {
+Status Posix::move_path(
+    const std::string& old_path, const std::string& new_path) {
   if (rename(old_path.c_str(), new_path.c_str()) != 0) {
     return LOG_STATUS(
         Status::IOError(std::string("Cannot move path: ") + strerror(errno)));
@@ -310,7 +311,7 @@ Status move_path(const std::string& old_path, const std::string& new_path) {
   return Status::Ok();
 }
 
-void purge_dots_from_path(std::string* path) {
+void Posix::purge_dots_from_path(std::string* path) {
   // Trivial case
   if (path == nullptr)
     return;
@@ -364,11 +365,14 @@ void purge_dots_from_path(std::string* path) {
     *path += std::string("/") + t;
 }
 
-Status read(
-    const std::string& path, uint64_t offset, void* buffer, uint64_t nbytes) {
+Status Posix::read(
+    const std::string& path,
+    uint64_t offset,
+    void* buffer,
+    uint64_t nbytes) const {
   // Checks
   uint64_t file_size;
-  RETURN_NOT_OK(posix::file_size(path, &file_size));
+  RETURN_NOT_OK(this->file_size(path, &file_size));
   if (offset + nbytes > file_size)
     return LOG_STATUS(
         Status::IOError("Cannot read from file; Read exceeds file size"));
@@ -403,12 +407,12 @@ Status read(
   return Status::Ok();
 }
 
-Status sync(const std::string& path) {
+Status Posix::sync(const std::string& path) {
   // Open file
   int fd = -1;
-  if (posix::is_dir(path))  // DIRECTORY
+  if (is_dir(path))  // DIRECTORY
     fd = open(path.c_str(), O_RDONLY, S_IRWXU);
-  else if (posix::is_file(path))  // FILE
+  else if (is_file(path))  // FILE
     fd = open(path.c_str(), O_WRONLY | O_APPEND | O_CREAT, S_IRWXU);
   else
     return Status::Ok();  // If file does not exist, exit
@@ -437,52 +441,100 @@ Status sync(const std::string& path) {
   return Status::Ok();
 }
 
-Status write(
+Status Posix::write(
     const std::string& path, const void* buffer, uint64_t buffer_size) {
-  // Open file
-  int fd = open(path.c_str(), O_WRONLY | O_APPEND | O_CREAT, S_IRWXU);
+  uint64_t file_offset = 0;
+  if (is_file(path) && !file_size(path, &file_offset).ok()) {
+    return LOG_STATUS(
+        Status::IOError(std::string("Cannot write to file '") + path));
+  }
+
+  // Open or create file.
+  int fd = open(path.c_str(), O_WRONLY | O_CREAT, S_IRWXU);
   if (fd == -1) {
     return LOG_STATUS(Status::IOError(
-        std::string("Cannot write to file '") + path + "'; " +
-        strerror(errno)));
+        std::string("Cannot open file '") + path + "'; " + strerror(errno)));
   }
 
-  // Append data to the file in batches of constants::max_write_bytes
-  // bytes at a time
-  uint64_t buffer_bytes_written = 0;
-  const char* buffer_bytes_ptr = static_cast<const char*>(buffer);
-  while (buffer_size > constants::max_write_bytes) {
-    uint64_t bytes_written = write_all(
-        fd,
-        buffer_bytes_ptr + buffer_bytes_written,
-        constants::max_write_bytes);
-    if (bytes_written != constants::max_write_bytes) {
-      return LOG_STATUS(Status::IOError(
-          std::string("Cannot write to file '") + path +
-          "'; File writing error"));
+  // Ensure that each thread is responsible for at least min_parallel_size
+  // bytes, and cap the number of parallel operations at the thread pool size.
+  uint64_t num_ops = std::min(
+      std::max(buffer_size / vfs_params_.min_parallel_size_, uint64_t(1)),
+      vfs_thread_pool_->num_threads());
+
+  if (num_ops == 1) {
+    if (!write_at(fd, file_offset, buffer, buffer_size).ok()) {
+      close(fd);
+      return LOG_STATUS(
+          Status::IOError(std::string("Cannot write to file '") + path));
     }
-    buffer_bytes_written += bytes_written;
-    buffer_size -= bytes_written;
-  }
-  uint64_t bytes_written =
-      write_all(fd, buffer_bytes_ptr + buffer_bytes_written, buffer_size);
-  if (bytes_written != buffer_size) {
-    return LOG_STATUS(Status::IOError(
-        std::string("Cannot write to file '") + path +
-        "'; File writing error"));
+  } else {
+    STATS_COUNTER_ADD(vfs_posix_write_num_parallelized, 1);
+    std::vector<std::future<Status>> results;
+    uint64_t thread_write_nbytes = utils::ceil(buffer_size, num_ops);
+
+    for (uint64_t i = 0; i < num_ops; i++) {
+      uint64_t begin = i * thread_write_nbytes,
+               end =
+                   std::min((i + 1) * thread_write_nbytes - 1, buffer_size - 1);
+      uint64_t thread_nbytes = end - begin + 1;
+      uint64_t thread_file_offset = file_offset + begin;
+      auto thread_buffer = reinterpret_cast<const char*>(buffer) + begin;
+      results.push_back(vfs_thread_pool_->enqueue(
+          [fd, thread_file_offset, thread_buffer, thread_nbytes]() {
+            return write_at(
+                fd, thread_file_offset, thread_buffer, thread_nbytes);
+          }));
+    }
+
+    bool all_ok = vfs_thread_pool_->wait_all(results);
+    if (!all_ok) {
+      close(fd);
+      return LOG_STATUS(
+          Status::IOError(std::string("Cannot write to file '") + path));
+    }
   }
 
-  // Close file
   if (close(fd) != 0) {
     return LOG_STATUS(Status::IOError(
         std::string("Cannot close file '") + path + "'; " + strerror(errno)));
   }
 
-  // Success
   return Status::Ok();
 }
 
-}  // namespace posix
+Status Posix::write_at(
+    int fd, uint64_t file_offset, const void* buffer, uint64_t buffer_size) {
+  // Append data to the file in batches of constants::max_write_bytes
+  // bytes at a time
+  uint64_t buffer_bytes_written = 0;
+  const char* buffer_bytes_ptr = static_cast<const char*>(buffer);
+  while (buffer_size > constants::max_write_bytes) {
+    uint64_t bytes_written = pwrite_all(
+        fd,
+        file_offset + buffer_bytes_written,
+        buffer_bytes_ptr + buffer_bytes_written,
+        constants::max_write_bytes);
+    if (bytes_written != constants::max_write_bytes) {
+      return LOG_STATUS(Status::IOError(
+          std::string("Cannot write to file; File writing error")));
+    }
+    buffer_bytes_written += bytes_written;
+    buffer_size -= bytes_written;
+  }
+  uint64_t bytes_written = pwrite_all(
+      fd,
+      file_offset + buffer_bytes_written,
+      buffer_bytes_ptr + buffer_bytes_written,
+      buffer_size);
+  if (bytes_written != buffer_size) {
+    return LOG_STATUS(Status::IOError(
+        std::string("Cannot write to file; File writing error")));
+  }
+
+  // Success
+  return Status::Ok();
+}
 
 }  // namespace sm
 }  // namespace tiledb